requirements.txt
aiogram==3.3.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
alembic==1.13.1
